Improved User Flairs
Border changed from Ridge to Outset for less noise on page.
User Flairs now have background colours that depend on the character
selected.
Fixed an alignment error with the User Flair text.
Removed Mod Flair colouring.
Added a text shadow to User Flair text so that the text is readable
regardless of the background colour.